pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "draftguru"
version = "0.1.0"
description = "FastAPI + SQLModel skeleton for a draft analytics app."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]==0.116.1",
    "sqlmodel==0.0.24",
    "pydantic-settings==2.10.1",
    "psycopg==3.2.9",
    "asyncpg==0.30.0",
    "alembic==1.16.5",
    "beautifulsoup4==4.12.3",
    "boto3",
    "google-genai",
    "numpy==2.1.3",
    "pandas==2.2.3",
    "scipy==1.14.1",
    "feedparser>=6.0.0",
    "resvg-py>=0.2.0",
    "Pillow>=10.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "httpx==0.28.1",
    "ruff==0.13.0",
    "mypy==1.18.1",
    "pre-commit==3.8.0",
    "pandas-stubs"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.mypy]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.cli.compute_metrics"
disable_error_code = "attr-defined,arg-type,call-overload,misc"

[tool.ruff]
exclude = ["alembic", "tests"]

[tool.ruff.lint]
extend-select = ["D"]
# Ignore missing docstring rules - only enforce formatting on existing docstrings
ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107"]

[tool.ruff.lint.pydocstyle]
convention = "google"
